+++ pages/root.md
@@ -1,0 +1,5 @@
+<|toggle|theme|>
+
+<center>
+<|navbar|>
+</center>
