+++ pages/root.py
@@ -1,0 +1,4 @@
+from taipy.gui import Markdown
+
+
+root_md = Markdown("pages/root.md")
